+++ doc/changes.rst
@@ -5,14 +5,15 @@ Change Log
 1.3.7 (unreleased)
 ------------------
 
+* Fix module processing problem for non-DESI Python packages.
+* Allow NERSC Modules root directory to be overridden in a configuration file.
+
 1.3.6 (2016-03-25)
 ------------------
 
 * Include funcfits in the documentation; added mk_fit_dict.
 * Improve coverage of funcfits.
 * Try to use a nicer Sphinx theme for documentation.
-* Fix module processing problem for non-DESI Python packages.
-* Allow NERSC Modules root directory to be overridden in a configuration file.
 
 1.3.5 (2016-03-15)
 ------------------