+++ doc/changes.rst
@@ -8,6 +8,7 @@ Change Log
 * Fix module processing problem for non-DESI Python packages.
 * Allow NERSC Modules root directory to be overridden in a configuration file.
 * Stats module was previously snuck in, but never documented.
+* Minor fixes for desiInstall bootstrap mode.
 
 1.3.6 (2016-03-25)
 ------------------